Refactored Delete Player
Delete a Player active in a Game will return 204 but the player will
not be deleted but set to inactive

+++ matchostat/matchostat-ejb/src/main/java/com/clowndata/service/GameServiceImpl.java
@@ -80,14 +80,18 @@ public class GameServiceImpl implements GameService {
 
         Player player = playerRepository.getPlayer(playerId);
 
-        List<Team> teams = gameRepository.getAllTeams();
+        if (player.isDeletable()) {
+            List<Team> teams = gameRepository.getAllTeams();
 
-        for (Team team : teams) {
-            if (team.getPlayers().contains(player)) {
-                team.getPlayers().remove(player);
+            for (Team team : teams) {
+                if (team.getPlayers().contains(player)) {
+                    team.getPlayers().remove(player);
+                }
             }
+            playerRepository.deletePlayer(playerId);
+        } else {
+            player.setActive(false);
         }
-        playerRepository.deletePlayer(playerId);
     }
 
     public GameSummary getGameSummary(Long gameId) {
